// src/reducers/actions/enum.ts
export enum CarActionTypesEnum {
    SET_ALL = 'SET_ALL',

    SET_ADDED_CARS = 'SET_ADDED_CARS',

    SET_EDITED_CARS = 'SET_EDITED_CARS',
    SET_REMOVED_CARS = 'SET_REMOVED_CARS',

    SET_NULL = 'SET_NULL',
    SET_CARS_BY_COMPANY_NAME = 'SET_CARS_BY_COMPANY_NAME',
    SET_CARS_BY_MODEL = 'SET_CARS_BY_MODEL',
    SET_CARS_BY_COLOR = 'SET_CARS_BY_COLOR',
    SET_CARS_BY_YEAR = 'SET_CARS_BY_YEAR',
    SET_CARS_BY_YEAR_LT = 'SET_CARS_BY_YEAR_LT',
    SET_CARS_BY_YEAR_GT = 'SET_CARS_BY_YEAR_GT',
}